requirements.txt
flask
flask-sqlalchemy
flask-migrate
python-dotenv
openai
pytest
pytest-flask
